src: gpu: intel: incl: fix implicit conversion warnings

+++ src/gpu/intel/include/math_utils.h
@@ -44,11 +44,12 @@ int __attribute__((overloadable)) rnd_up(int a, unsigned int b) {
     return div_up(a, b) * b;
 }
 
-int __attribute__((overloadable)) rnd_up(unsigned int a, unsigned int b) {
+unsigned int __attribute__((overloadable)) rnd_up(
+        unsigned int a, unsigned int b) {
     return div_up(a, b) * b;
 }
 
-int __attribute__((overloadable)) rnd_up(long a, unsigned int b) {
+long __attribute__((overloadable)) rnd_up(long a, unsigned int b) {
     return div_up(a, b) * b;
 }
 
@@ -56,11 +57,12 @@ int __attribute__((overloadable)) rnd_down(int a, unsigned int b) {
     return (a / b) * b;
 }
 
-int __attribute__((overloadable)) rnd_down(unsigned int a, unsigned int b) {
+unsigned int __attribute__((overloadable)) rnd_down(
+        unsigned int a, unsigned int b) {
     return (a / b) * b;
 }
 
-int __attribute__((overloadable)) rnd_down(long a, unsigned int b) {
+long __attribute__((overloadable)) rnd_down(long a, unsigned int b) {
     return (a / b) * b;
 }
